Rebase with password error states
