+++ app/src/main/java/org/mozilla/fenix/settings/logins/SavedLoginsView.kt
@@ -73,7 +73,7 @@ class SavedLoginsView(
  * Interactor for the saved logins screen
  */
 class SavedLoginsInteractor(
-    private val savedLoginsController: DefaultSavedLoginsController,
+    private val savedLoginsController: SavedLoginsController,
     private val itemClicked: (SavedLogin) -> Unit,
     private val learnMore: () -> Unit
 ) {
